Update automation to assign action items to workflow actor (#2333)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1206628689034179/f

Description:
This change updates release automation to mention the person running the workflow in log messages
and assign action items to them, instead of the release task assignee. If the actor ID is not recognized
or missing (such as when a workflow is run on schedule), it falls back to the release task assignee.