Alessandro/address bookmarks list fix menu (#2352)
Task/Issue URL: https://app.asana.com/0/0/1206762730363836/f

**Description**:
I found an issue while doing a smoke test for the Ship Review. The first time a bookmark/folder is added from the bookmarks shortcut panel and we open the menu from the “…” menu button, the menu is disabled. The same
is not true if we open the menu by right-clicking on the item.