Ensure that root bookmark folder comparison works when one folder has… (#2354)
Task/Issue URL: https://app.asana.com/0/0/1206762730363884/f

**Description**:
In some cases `BookmarkFolder`s that are added to the root folder have `parentFolderUUID` equal to `nil` in some cases they have it equal to the root identifier `bookmarks_root`. Ensure that equality is satisfied when this happens.
This bug started when we changed hashable/equatable.