Add a placeholder en-US translation to fix Monterey and Big Sur (#2396)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1206826260134772/f

Description:
This change adds English (US) to SyncUI's string catalog to fix English language on macOS
Big Sur and Monterey. One string is translated so that the entry for en_US appears in the string catalog,
which makes the OS recognize that English language translation is present in the SyncUI bundle.
Other (non-translated) strings for en_US are taken from the default translation (English) that is compiled
into the bundle.