Prevent Dismissal of Autofill UI When Window Loses Focus  (#2393)
Task/Issue URL: https://app.asana.com/0/276630244458377/1204870766898373/f

Description:
Currently, if we open the Autofill dialog and then tap out of the browser, the autofill window immediately dismisses. This is not a great UX. This change ensures that the Autofill dialog remains open until the user explicitly decides to close it.
Note that the user can close the autofill dialog by selecting anywhere inside the DDG browser, but outside the Autofill dialog.