Display error messaging for cancelled subscriptions (#2394)
Task/Issue URL: https://app.asana.com/0/72649045549333/1205438842252963/f

Description:

When we detect that entitlements have expire AND Netp is enabled:

Disable NetP
Show revoked notification and make sure that Netp disabled notification is NOT shown.
On opening the status view we show the revoked dialog.
When we get 403 response from /register (or any controller APIs), we handle the expired entitlement (refer to first bullet point in this section). (Handled by the iOS equivalent project
When rekey is attempted with invalid entitlements, we will get 403. (Handled by the iOS equivalent project
If user attempts to enable Netp with invalid entitlements, we will get 403. (Handled by the iOS equivalent project
We do a periodic check every 20 minutes calling the subscription API to get entitlements (Handled by the NetworkProtectionEntitlementsMonitor implemented in the iOS equivalent project )
If check succeeds and user has entitlement, we do nothing.
If check succeeds and user has no entitlement, we handle the expired entitlement (refer to first bullet point in this section).
If check fails (no internet, BE failure, tunnel failure), we do nothing.
If Netp is disabled, we stop this check.
On top of this, we also added an additional check on the Settings page.

To correctly show the Netp entry, we check for entitlement through subscription API whenever the Settings page is foregrounded.
If check succeeds and user has entitlement, we show the Netp settings entry.
If check succeeds and user has no entitlement, we handle the expired entitlement (refer to first bullet point in this section) and also hide Netp settings entry.
If check fails (no internet, be failure, tunnel failure), we retain whatever the previous visibility of the Settings entry if available, else hide entry.