Change how subscriptionFeatureAvailability.isFeatureAvailable is used (#2489)
Task/Issue URL:
https://app.asana.com/0/1200019156869587/1206905704487790/f
Tech Design URL:
CC:

**Description**:
Get rid of `subscriptionFeatureAvailability` from the AppDelegate and
instead instantiate it locally where needed.


**Steps to test this PR**:
Smoke test if Privacy Pro is hidden according to the
`isFeatureAvailable` check (e.g. when internal state is on/off)

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)