Expand dbp error handling iteration 2: further up the stack + pixels (#2484)
Task/Issue URL:
https://app.asana.com/0/1199230911884351/1206910398550646/f
Tech Design URL:
CC:

**Description**: 
Adds additional error handling higher up the stack, as well as firing
pixels for DBP errors.
This should not be merged yet because privacy triage isn't complete

Note that this depends on iteration 1, and is set with that as the base

**Steps to test this PR**:
The main thing to test is error pixels are fired as expected. There are
two general cases to to test:
1. The secure vault error pixels, both from the background agent and
from the main app
2. The new "generalError" pixel (again both from the background agent
and from the main app). Pay special attention to making sure it always
has at least the following params: error code, error domain, and the
origin function. Special care is needed here with the origin function,
because if one uses pixel kit debug events it clobbers the original
params

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)

---------

Co-authored-by: Dominik Kapusta <dkapusta@duckduckgo.com>
Co-authored-by: Fernando Bunn <bunn@idevzilla.com>
Co-authored-by: Dax the Duck <dax@duckduckgo.com>
Co-authored-by: Dax Mobile <44842493+daxmobile@users.noreply.github.com>
Co-authored-by: Diego Rey Mendez <diegoreymendez@users.noreply.github.com>
Co-authored-by: Pete Smith <peadar81@gmail.com>
Co-authored-by: Sabrina Tardio <44158575+SabrinaTardio@users.noreply.github.com>
Co-authored-by: Anh Do <18567+quanganhdo@users.noreply.github.com>
Co-authored-by: Juan Manuel Pereira <juanmanuel.pereira1@gmail.com>
Co-authored-by: Christopher Brind <brindy@duckduckgo.com>
Co-authored-by: Michal Smaga <miasma13@gmail.com>
Co-authored-by: bwaresiak <bartek@duckduckgo.com>
Co-authored-by: Mariusz Śpiewak <dus7@users.noreply.github.com>
Co-authored-by: Daniel Bernal <dbernal@duckduckgo.com>
Co-authored-by: muodov <muodov@users.noreply.github.com>
Co-authored-by: Sam Symons <sam@samsymons.com>
Co-authored-by: Alessandro Boron <aboron@duckduckgo.com>
Co-authored-by: Federico Cappelli <fcappelli@duckduckgo.com>
Co-authored-by: Tom Strba <57389842+tomasstrba@users.noreply.github.com>
Co-authored-by: Graeme Arthur <graeme@graemearthur.com>
Co-authored-by: amddg44 <amallon@duckduckgo.com>
Co-authored-by: Brad Slayter <SlayterDev@users.noreply.github.com>
Co-authored-by: Halle <378795+Halle@users.noreply.github.com>
Co-authored-by: Diego Rey Mendez <dreymendez@duckduckgo.com>