Adds a series of UI tests for Bookmarks Bar visibility
Task/Issue URL:
https://app.asana.com/0/1199230911884351/1205717021705368/f
Tech Design URL:
CC:

**Description**: Adds a series of UI tests for Bookmarks Bar visibility

**Steps to test this PR**:
1. Open the scheme **UI Tests**
2. Navigate to the test pane
3. Run BookmarksBarTests

Note: this is the first PR where the scheme builds `review`, so it
should no longer be necessary to comment out
`app.launchEnvironment["UITEST_MODE"] = "1"`. However, if this builds a
`review` build that is treated as a first run on your system during
every test, please build and run the scheme once instead of running its
tests, and go through the first run steps once before running the tests
(I will check in with @kshann about how to deal with this in CI, or if
it is necessary to do so).