Update Neighbor Report (#2542)
Task/Issue URL: https://app.asana.com/0/0/1206998831911062/f
Tech Design URL:
CC:

**Description**: Sometimes Neighbor Report fails to load the page after
the opt out form (where buttons need to be clicked). When this happens,
we were reporting errors on the "click" action which was confusing for
those needing to debug. This PR adds an expectation to check and make
sure we're looking at the correct page before clicking buttons.