Fixes the VPN restarting logic on update (#2545)
Task/Issue URL:
https://app.asana.com/0/1199230911884351/1206999858530015/f

## Description:

I noticed that we're getting increased number of failures when a new
version of our app is released. After researching this I found that the
mechanism we use for restarting the VPN can sometimes fail, but even if
it doesn't fail it always fires failure pixels.

This is likely caused by the fact that not disabling on-demand makes the
connection try to start while we're stopping it and starting it
ourselves.

We'll now go back to disabling on-demand, but we'll now enable it
optimistically right after starting the VPN again.