Listen for resign active events and dismiss the VPN popover.
Users occasionally report confusion about the VPN popover being greyed out, even though it’s active. This is because the popover stays open when the app loses focus, so this commit changes it to dismiss when the app loses focus.

However, we want to keep the popover open if the user is onboarding, because otherwise the popover will vanish whenever they get prompted to do something.