Add additonal error handling to DBP database and related classes (#2384)
Task/Issue URL: https://app.asana.com/0/0/1206814903024662/f
Tech Design URL:
CC:

**Description**:
Adds additional error handling mostly to the DB classes but also others.
This requires updating quite a lot of classes to take this into account.
Also makes the DBP tests have the same swift lint rules as the main unit
tests

Things this does not do:
- Necessarily arrive at the final desired error handling for each class
updated. Some errors are ignored with inert do/catches to allow the PR
to be delivered without updating the entire codebase.
- Add pixels (since they require a privacy triage) 

Also worth noting that sometimes this error handling introduces subtle
behavior changes, where if a function uses a function that throws, it
will now be rethrown, rather than the function plowing ahead regardless.
I don't think there are any cases where this is an undesirable change,
but worth calling out in case it turns out we were having non blocking
errors that now suddenly become blocking. Long term I consider this to
be a feature and part of the point.

**Steps to test this PR**:
1. Obviously try the core functionality of DBP, but outside of that I
don't think manual testing is the way to go with this
2. Now the other PRs are merged into this, there are some other things
to test:

- That error pixels are fired as expected. There are two general cases
to to test:

The secure vault error pixels, both from the background agent and from
the main app
The new "generalError" pixel (again both from the background agent and
from the main app). Pay special attention to making sure it always has
at least the following params: error code, error domain, and the origin
function. Special care is needed here with the origin function, because
if one uses pixel kit debug events it clobbers the original params

- Simulate throwing various kinds of errors and make sure the processor
receives them from the collection operations and fires the original
pixel

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)

---------

Co-authored-by: Dominik Kapusta <dkapusta@duckduckgo.com>
Co-authored-by: Fernando Bunn <bunn@idevzilla.com>
Co-authored-by: Dax the Duck <dax@duckduckgo.com>
Co-authored-by: Dax Mobile <44842493+daxmobile@users.noreply.github.com>
Co-authored-by: Diego Rey Mendez <diegoreymendez@users.noreply.github.com>
Co-authored-by: Pete Smith <peadar81@gmail.com>
Co-authored-by: Sabrina Tardio <44158575+SabrinaTardio@users.noreply.github.com>
Co-authored-by: Anh Do <18567+quanganhdo@users.noreply.github.com>
Co-authored-by: Juan Manuel Pereira <juanmanuel.pereira1@gmail.com>
Co-authored-by: Christopher Brind <brindy@duckduckgo.com>
Co-authored-by: Michal Smaga <miasma13@gmail.com>
Co-authored-by: bwaresiak <bartek@duckduckgo.com>
Co-authored-by: Mariusz Śpiewak <dus7@users.noreply.github.com>
Co-authored-by: Daniel Bernal <dbernal@duckduckgo.com>
Co-authored-by: muodov <muodov@users.noreply.github.com>
Co-authored-by: Sam Symons <sam@samsymons.com>
Co-authored-by: Alessandro Boron <aboron@duckduckgo.com>
Co-authored-by: Federico Cappelli <fcappelli@duckduckgo.com>
Co-authored-by: Tom Strba <57389842+tomasstrba@users.noreply.github.com>
Co-authored-by: Graeme Arthur <graeme@graemearthur.com>
Co-authored-by: amddg44 <amallon@duckduckgo.com>
Co-authored-by: Brad Slayter <SlayterDev@users.noreply.github.com>
Co-authored-by: Halle <378795+Halle@users.noreply.github.com>
Co-authored-by: Diego Rey Mendez <dreymendez@duckduckgo.com>