Adds a series of UI tests for Address Bar Keyboard
Task/Issue URL:
https://app.asana.com/0/1199230911884351/1207045927838862/f
Tech Design URL:
CC:

**Description**: Adds a series of UI tests for Address Bar Keyboard
Shortcuts

**Steps to test this PR**:
1. Open the scheme **UI Tests**
2. Navigate to the test pane
3. Run `AddressBarKeyboardShortcutsTests`

**UI Tests general guidelines for everyone**: 

* It isn't possible to multitask while running UI tests on your local
machine, because your mousing/interface usage or the unknown focus
events of other in-use applications can change, slow down, or intercept
screen interactions that the tests depend on. Multitasking includes
being on calls, and includes work in a different space. Tedious as it
is, the UI tests are the only thing you can do with your computer when
you test whether UI tests pass, because that is the only circumstance
that `XCUIAutomation` effectively supports locally (this makes sense if
you consider how closed the simulator environment where this API gets
the most development is, by contrast with an engineer's development
computer). This approach keeps the debug loop fast and the tests free of
workaround buildup.
* Due to app-specific text field focus issues related to waiting for
focus across two monitors simultaneously, we've decided to review the
tests on a single monitor.
* English is the currently supported language for UI tests
* The tests have been tested on multiple systems before the PR is
opened. If you experience a failure when testing locally, in order to
confirm that it isn't a false negative failure, please take the
following steps before reporting it:
1. If you weren't watching when it happened (probably not, it's pretty
tedious) please re-run the failed case and observe the failure. This
does a few things: it means the person most acquainted with the system
where the test failed can describe what they observed and knows what the
test tries to do, which will promote a faster and more minimal solution,
and it is a way of ruling out accidental multitasking, which can happen
if a call comes in or a long GUI process elsewhere concludes. If a
failure recurs when you observe the test case, please report it.
2. When reporting it, please send the `xcresult` bundle, even if it
seems like the same failure as a previous failure (it may be subtly
different). Since the `xcresult` bundle will include a screen recording
or screenshots, for your privacy, please consider the visibility of
things in your system you may not want to send to a contractor when you
are running the tests.

Thank you very much for your help!