Remove "Thank You" prompts (#2659)
This only removes the calls to show the prompts themselves. All of the
UI classes and copy will be cleaned up on the main branch, this is done
to keep the release branch change as small as possible.

Task/Issue URL:
https://app.asana.com/0/1199230911884351/1207105265255740/f
Tech Design URL:
CC:

**Description**:

This PR removes the Thank You prompts that are presented to waitlist
users.

This is being done because the code is being taken down on the Stripe
side next week, so we should no longer show the prompt to waitlist users
who haven't yet upgraded.