Support close tabs to left and move close actions to submenu (#2664)
Task/Issue URL:
https://app.asana.com/0/1199230911884351/1207117125840765/f
Tech Design URL:
CC: @ayoy 

**Description**:
This PR adds the option to close all tabs to the left when right
clicking a tab. It also moves this action as well as "close to right"
and "close other tabs" to a submenu in the tab context menu.

**Steps to test this PR**:
1. Open some tabs
2. Right click a tab, ensure "Close Other Tabs" opens a sub menu
3. Clicking "Close Tabs to the Left" should close all tabs to the left
of the clicked tab except pinned tabs
4. "Close to the right" and "Close all other tabs" should work as
expected.

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)

---------

Co-authored-by: Dominik Kapusta <dkapusta@duckduckgo.com>