Automatically clear data upon quitting (#2600)
Task/Issue URL:
https://app.asana.com/0/1177771139624306/1205062321200340/f

**Description**:
Adding Burn on Quit feature for macOS. Upon standard exit, data are
cleared and the fire animation is presented. If macOS is restarting, the
app delays the restart until all data have been cleared. In the event of
an unexpected termination, such as a crash or force quit, data clearing
is handled at the next startup.