Send pixels when showing notifications (#2681)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1207142220574385/f
Tech Design URL:
CC:

Description:

This PR sends pixels when we display notifications to the user. We're doing this temporarily, so that we can verify that notifications are not showing in massive rates for some users.

This PR has a larger diff than expected since the notifications target was never set up to send pixels until now.