Egress Server Failure Detection & Recovery (#2657)
Task/Issue URL:
https://app.asana.com/0/72649045549333/1206669340827392/f
BSK PR: https://github.com/duckduckgo/BrowserServicesKit/pull/786

**Description**:
- Use the WireGuard handshakes to determine when a connection fails by
checking that the last handshake time is less than 300 seconds (just
over two handshakes).
- When a failure is detected the client will call /register with server
set to the name of the current (failed) server and mode set to
failureRecovery. If the server is not in the "active" status on the
controller a new registration will be returned otherwise the previous
data will be returned (this data should be checked against the current
tunnel data including allowedIPs).

**Steps to test this PR**:
1. Launch the VPN and connect 
2. Block UDP traffic by following these instructions:
https://app.asana.com/0/0/1207029295107667/f
3. Open the Console.app and filter by Category: `Network Protection`.
The failure recovery logs are indicated with a 🟢 so you can also look
for them.
4. After around 5 minutes you will see logs indicating that a register
request has been made.
5. It will error with no recovery required as there is not a real
problem.

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)