Merge pir stabilization feature branch into main (#2789)
Task/Issue URL:
https://app.asana.com/0/1199230911884351/1207338264132623/f
Tech Design URL:
https://app.asana.com/0/481882893211075/1207174884557414/f
CC:

**Description**:
This PR has the effect of mergeing four different PRs into main:
https://github.com/duckduckgo/macos-browser/pull/2777
https://github.com/duckduckgo/macos-browser/pull/2758
https://github.com/duckduckgo/macos-browser/pull/2757
https://github.com/duckduckgo/macos-browser/pull/2743
This covers significant changes to the XPC interface and how the main
app uses it, the background manager, the scheduler, and the processor.
See individual PRs for details.

There's no code changes here that haven't been reviewed separately as
part of those PRs, so it's up to you how you want to review it code
wise. The more important step at this stage is manual testing.

**Steps to test this PR**:
1. Because this is a fairly significant set of changes, we need to be as
thorough as we can in testing DBP generally works and really try to
break it. Because of this, I'm keen that anyone testing things of their
own ways to break it before trying the steps below.
2. Edit/add profile data, and before scans finish, edit it again. Scans
should start afresh, and you should see an interrupted pixel fire. Check
that you don't get a scans completed notification
3. Edit/add profile data, and before scans finish, close the app, and
reopen it. The same set of manual scans should still continue, and a
blocked pixel should fire.
4. With profile data already existing but initial scans having finished,
close the app and reopen it. Scheduled scans should run.
5. With profile data already existing, and with the app closed, launch
the background agent, scheduled scans should run
6. Edit/add profile data, put the laptop to sleep/lock it/restart it,
check that scans continue as expected.
7. Check that the background agent activity lines up with the UI, I.e.
the progress bar and when we sent the scans completed notification is
accurate

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)

---------

Co-authored-by: Pete Smith <peadar81@gmail.com>