Check DBP Prerequisites in App and Disable Login Item If Necessary (#2850)
Task/Issue URL:
https://app.asana.com/0/1206488453854252/1207501562611619/f

**Description**:
This PR introduces changes to check DBP prerequisites (user is
authenticated and has valid entitlements) from the App when the app
becomes active. When prerequisites are not satisfied, the app disables
the login item. This is to prevent a scenario where the background agent
checks prerequisites, exits due to failed prerequisites, and then is
automatically re-started by the login item.