Fixes an issue reporting the last extension version run (#2960)
Task/Issue URL:
https://app.asana.com/0/1207603085593419/1207785443093297/f

## Description

There's an issue where metadata keeps reporting the last extension
version run as "unknown".

This issue happens because we always overwrite the last extension
version run in the metadata, but it's only provided to us on update. We
should change the logic to only change the last extension version when
it's present and ignore it otherwise.