Allow activating subscription for internal users via debug menu (#3003)
Task/Issue URL:
https://app.asana.com/0/72649045549333/1207811184536020/f

**Description**:
Add to debug menu option to start the subscription activation flow. This
option is to allow access to the flow even when the purchase is
disallowed in given region (App Store build for non-US user).

Additional changes are renaming and cleanup of obsolete items in the
"Subscription" debug menu.

**Steps to test this PR**:
1. Remove subscription if any present.
2. Open menu select: Debug -> Subscription -> I Have a Subscription
3. A new tab should open for activating subscription via email
4. Activate subscription via email OTP flow

Test the above for the setup where purchase platform is set to App Store
and no subscription products are available so there are no purchase
entry points shown in the app. This can be achieved by altering
MAIN_BUNDLE_IDENTIFIER_PREFIX for a given build (so it does not matches
one we have subscriptions defined in the App Store) and after launching
the build ensuring that the purchase environment is set to App Store).
After these changes the "Privacy Pro" should not be visible in the more
options menu not in the settings but the activation flow from the debug
menu should work.

**Definition of Done**:

* [ ] Does this PR satisfy our [Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f)?

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)