bump CSS for new duckplayer (#3051)
Task/Issue URL:
https://app.asana.com/0/1201141132935289/1207941543404278/f
Tech Design URL:
CC:

**Description**:

As noted in Asana, this is about bringing the new Duck Player codebase
to Windows/macOS where we can verify how it's working in a known
settings.

**there should be no changes to how duck player works** - This PR is
successful if DuckPlayer operates just like the production one does.

NOTE: the CSS code is *not* released yet. The purpose of this PR +
testing is to help me validate that it's ready for release.

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

**Steps to test this PR**:
1. Keyboard focus: Verify that the Youtube embed has initial keyboard
focus - this can be checked by navigating to duckplayer and pressing
spacebar once the video has loaded
2. PIP: verify that Picture in Picture works as expected (you can see
the icon, and click it)
3. Error screen: verify that a url such as 'duck://player/123' shows an
error screen instead of the YouTube embed
4. Settings: verify that clicking the settings cog opens settings as
expected
5. Watch In YouTube: verify that clicking 'watch in youtube' works as
expected

**Definition of Done**:

* [ ] Does this PR satisfy our [Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f)?

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)

Co-authored-by: Shane Osbourne <sosbourne@duckduckgo.com>