Fix bookmarks sort UI tests (#3296)
Task/Issue URL: https://app.asana.com/0/1204006570077678/1208328455174895/f

Description:
This fixes the bookmark sort UI tests. I introduced some changes where I broke the tests (#3253); luckily, the tests I helped create found the issue!

The main problem is that now we disable sort and search when we do not have bookmarks. UI tests didn’t account for that, so we tried accessing the sort button without bookmarks, which is now impossible.

I fixed the issue and added UI tests to cover the disable states of the sort button when there are no bookmarks.