For third party requests differentiate if they are affiliated with first party (#3343)
Task/Issue URL:
https://app.asana.com/0/414709148257752/1208376794617030/f

**Description**:
Fixed in BSK.
When loading a website and some of the allowed third party requests may
not be recognized as trackers but by their URL they may belong to the
same entity as the website. For that cases we should mark their state as
.`allowed(reason: .ownedByFirstParty)` (instead of `.allowed(reason:
.otherThirdPartyRequest)`).

**Steps to test this PR**:
See parent task for testing steps.

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

**Definition of Done**:

* [ ] Does this PR satisfy our [Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f)?

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)