Includes the VPN proxy in all builds (#3393)
Task/Issue URL: https://app.asana.com/0/1207603085593419/1208543229974056/f

Description

The VPN proxy has been embedded a long time ago into the app and recently it was enabled for the domain exclusions feature. Unfortunately for our App Store builds, the proxy was only included in DEBUG builds.

This was done because Apple usually rejects builds that include components that cannot be used, so we needed to keep it out of release builds until it was used.

During development this was missed as the DEBUG builds include the proxy.