Add NTP Customization Settings onboarding popover (#3428)
Task/Issue URL: https://app.asana.com/0/0/1208583505730578/f

Description:
This change adds a popover that is displayed once for every user over the 'Customize' button on the NTP.
It's not a part of the experiment nor it's related to the search bar project. We're adding the onboarding to
increase engagement with NTP customization settings.